remove NODE_ENV and use separated scripts
