--- bin/server.js
@@ -1,46 +1,0 @@
-require('dotenv').config()
-const express = require('express')
-const webpack = require('webpack')
-const webpackDevMiddleware = require('webpack-dev-middleware')
-const webpackHotMiddleware = require('webpack-hot-middleware')
-const webpackHotServerMiddleware = require('webpack-hot-server-middleware')
-const app = require('../src/server').app
-
-const clientConfig = require('../webpack/client.dev')
-const serverConfig = require('../webpack/server.dev')
-const clientConfigProd = require('../webpack/client.prod')
-const serverConfigProd = require('../webpack/server.prod')
-
-const publicPath = clientConfig.output.publicPath
-const outputPath = clientConfig.output.path
-
-let isBuilt = false
-
-function done() {
-  if (!isBuilt) {
-    app.listen(process.env.APP_PORT, () => {
-      isBuilt = true
-      console.log(`Listening @ http://localhost:${process.env.APP_PORT}/`)
-    })
-  }
-}
-
-if (process.env.NODE_ENV === 'development') {
-  const compiler = webpack([clientConfig, serverConfig])
-  const clientCompiler = compiler.compilers[0]
-  const options = { publicPath, stats: { colors: true } }
-
-  app.use(webpackDevMiddleware(compiler, options))
-  app.use(webpackHotMiddleware(clientCompiler))
-  app.use(webpackHotServerMiddleware(compiler))
-
-  compiler.plugin('done', done)
-} else {
-  const clientStats = require('../buildClient/stats.json')
-  const serverRender = require('../buildServer/main.js').default
-
-  app.use(publicPath, express.static(outputPath))
-  app.use(serverRender({ clientStats }))
-
-  done()
-}
